add a general dart command
