+++ mono_repo/pubspec.yaml
@@ -2,7 +2,7 @@ name: mono_repo
 description: >-
   CLI tools to make it easier to manage a single source repository containing
   multiple Dart packages.
-version: 6.5.7
+version: 6.6.0
 repository: https://github.com/google/mono_repo.dart
 
 topics:
@@ -10,7 +10,7 @@ topics:
   - repository-management
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: ^3.0.0
 
 dependencies:
   args: ^2.0.0
